+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Minor: Modified help message for `flag-handler`
 - Minor: Moved `formatTime` and `getAccountRegion` to `utils`
 - Minor: Refactored notification codes check using optional chaining
+- Bugfix: Fixed bug where notification kept being send even skipCheck is set to false (#10)
 
 ## 6/17/2023
 - Major: Telegram users now can use commands to check stamina, and check expedition status by using `/stamina` and `/expedition` respectively.